+++ tests/PreflightTest.php
@@ -13,6 +13,18 @@ class PreflightTest extends TestCase
         $this->assertEquals(200, $crawler->getStatusCode());
     }
 
+    public function testAllowWildcardOriginAllowed()
+    {
+        config(['cors.allowedOrigins' => ['*.laravel.com']]);
+
+        $crawler = $this->call('OPTIONS', 'api/ping', [], [], [], [
+            'HTTP_ORIGIN' => 'blog.laravel.com',
+            'HTTP_ACCESS_CONTROL_REQUEST_METHOD' => 'POST',
+        ]);
+        $this->assertEquals('blog.laravel.com', $crawler->headers->get('Access-Control-Allow-Origin'));
+        $this->assertEquals(200, $crawler->getStatusCode());
+    }
+
     public function testAllowOriginNotAllowed()
     {
         $crawler = $this->call('OPTIONS', 'api/ping', [], [], [], [