+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## v0.11.0 (2017-12-xx)
+### Breaking changes
+ - The wildcard matcher is changed. You can use `allowedOriginPatterns` for your own patterns, 
+ or simple wildcards in the normal origins. Eg. `*.laravel.com` should still work.
+
 ## v0.9.0 (2016-03-2017)
 ### Breaking changes
  - The `cors` alias is no longer added by default. Use the full class or add the alias yourself.